chore(lightgbm): move SchemaFieldsUtil to lightgbm package
